Great Barrier Reef Features

Coastal features within and adjacent to the Great Barrier Reef World Heritage area (such as islands, rocks and the Queensland mainland). - Major coral cay features - Major and other coral reef structures (as defined by the reef shoal edge) and tidal, drying or emergent reef areas within the Great Barrier Reef World Heritage Area.

Format

A Simple Features data.frame with 5,363 rows and 21 variables:

OBJECTID:

SORT_GBR_I:

LABEL_ID:

SUB_NO:

CODE:

UNIQUE_ID:

FEATURE_C:

GBR_NAME:

FEAT_NAME:

QLD_NAME:

X_LABEL:

GBR_ID:

LOC_NAME_S:

LOC_NAME_L:

X_COORD:

Y_COORD:

Area_HA:

GlobalID:

Shape_STAr:

Shape_STLe:

geometry:

Details

For full Metadata see the original GBRMPA Geoportal page info for the Great Barrier Reef Features.

Data kindly provided by: Great Barrier Reef Marine Park Authority Data Administrator Spatial Data Centre PO Box 1379 Townsville, QLD 4810 Australia gis@gbrmpa.gov.au

The Great Barrier Reef Marine Park Authority requests attribution in the following manner: Copyright (c) Great Barrier Reef Marine Park Authority 2020. This product is released under the Creative Commons Attribution 4.0 International Licence. Updated data available at http://www.gbrmpa.gov.au/geoportal/. Under CC BY 4.0, users are free to share and adapt the data under the following conditions:

They must give appropriate credit

Provide a link to the license

Indicate if any changes were made, and

They may not apply any additional restrictions that legally restrict others from doing anything the license permits.