Great Barrier Reef Marine Park Zoning

The Great Barrier Reef Marine Park Act 1975 (the Act) provides for the establishment, control, care and development of the Great Barrier Reef Marine Park (the Marine Park). The Great Barrier Reef Marine Park Authority (the Authority) is responsible for the management of the Marine Park. The Great Barrier Reef Marine Park Zoning Plan 2003 is the primary planning instrument for the conservation and management of the Marine Park. In having regard to the objects set out in subsection 32 (7) of the Act, this Zoning Plan takes account of the world heritage values of the Marine Park and the principles of ecologically sustainable developments. This Zoning Plan aims, in conjunction with other management mechanisms, to protect and conserve the biodiversity of the Great Barrier Reef ecosystem within a network of highly protected zones, while providing opportunities for the ecologically sustainable use of, and access to, the Great Barrier Reef Region by current and future generations. In addition to the protection of representative areas of biodiversity, this Zoning Plan also provides for the protection of other areas of high conservation value by assigning protective zoning to a range of habitats such as coral reefs, sponge beds, seagrass beds and deep water areas, as well as important dugong habitats and other special or unique sites. The Marine Park is managed as a multiple use area. This means that, while enhancing the conservation of the Marine Park, this Zoning Plan also provides for a range of recreational, commercial and research opportunities, and for the continuation of traditional activities. This Zoning Plan expressly acknowledges the rights and interests of indigenous Australians in the Marine Park by providing for the management of the traditional use of marine resources, including traditional hunting, in accordance with Aboriginal and Torres Strait Islander custom and tradition. The contributions of scientific research to the management and understanding of the Marine Park are also acknowledged in this Zoning Plan, which provides for the management of research in the Marine Park, including a system of Scientific Research Zones to facilitate research around scientific research stations.(Extract from "Preface A, Introduction; Great Barrier Reef Marine Park Zoning Plan 2003).

Format

A Simple Features data.frame with 756 rows and 26 variables:

OBJECTID:

NAME:

IUCN:

NRS_PA:

GAZ_AREA:

GIS_AREA:

RES_NO:

GAZ_DATE:

LATEST_GAZ:

STATE:

AUTHORITY:

ENVIRON:

CODE:

X_COORD:

Y_COORD:

ALT_ZONE:

TYPE:

AREA_TYPE:

LOC_NAME_S:

PERMIT_DES:

GLOBALID:

GROUP_ID:

STD_PERMIT:

Shape_area:

Shape_len:

geometry:

Details

For full Metadata see the original GBRMPA Geoportal page info for the Great Barrier Reef Marine Park Zoning.

Data kindly provided by: Great Barrier Reef Marine Park Authority Data Administrator Spatial Data Centre PO Box 1379 Townsville, QLD 4810 Australia gis@gbrmpa.gov.au

The Great Barrier Reef Marine Park Authority requests attribution in the following manner: Copyright (c) Great Barrier Reef Marine Park Authority 2020. This product is released under the Creative Commons Attribution 4.0 International Licence. Updated data available at http://www.gbrmpa.gov.au/geoportal/. Under CC BY 4.0, users are free to share and adapt the data under the following conditions:

They must give appropriate credit

Provide a link to the license

Indicate if any changes were made, and

They may not apply any additional restrictions that legally restrict others from doing anything the license permits.