Marine Water Bodies within the Great Barrier Reef

The GBRMPA delineation as published in the guidelines are: Enclosed Coastal (Queensland Water Quality Guidelines); Open Coastal (AIMS description of coastal water body relative distance from Research Publication no. 89); Midshelf (AIMS description of inner shelf water body relative distance from Research Publication no. 89); and Offshsore (AIMS description of relative distance from Research Publication no. 89) Coral Sea taken from DEWHA definition of current Coral Sea Conservation Park.

Format

A Simple Features data.frame with 485 rows and 16 variables:

typesource:

MarineWate:

UniqueID:

chl_annnua:

chl_50th:

chl_80th:

TSS_annual:

TSS_50th:

TSS_80th:

chl_source:

TSS_source:

Lat_DD:

Lon_DD:

Shape_STAr:

Shape_STLe:

geometry:

Details

Data kindly provided by: Great Barrier Reef Marine Park Authority Data Administrator Spatial Data Centre PO Box 1379 Townsville, QLD 4810 Australia gis@gbrmpa.gov.au

The Great Barrier Reef Marine Park Authority requests attribution in the following manner: Copyright (c) Great Barrier Reef Marine Park Authority 2020. This product is released under the Creative Commons Attribution 4.0 International Licence. Updated data available at http://www.gbrmpa.gov.au/geoportal/. Under CC BY 4.0, users are free to share and adapt the data under the following conditions:

They must give appropriate credit

Provide a link to the license

Indicate if any changes were made, and

They may not apply any additional restrictions that legally restrict others from doing anything the license permits.